+++ db/migrations/20160616094519_drop_message_from_external_errors.rb
@@ -1,0 +1,7 @@
+Hanami::Model.migration do
+  change do
+    alter_table :external_errors do
+      drop_column :message
+    end
+  end
+end