+++ spec/lib/concierge/flows/external_error_creation_spec.rb
@@ -7,7 +7,6 @@ RSpec.describe Concierge::Flows::ExternalErrorCreation do
       supplier:    "SupplierA",
       code:        "http_error",
       context:     { type: "network_failure" },
-      message:     "Network Failure",
       happened_at: Time.now
     }
   }
@@ -15,7 +14,7 @@ RSpec.describe Concierge::Flows::ExternalErrorCreation do
   describe "#perform" do
     subject { described_class.new(parameters) }
 
-    [:operation, :supplier, :code, :context, :message, :happened_at].each do |required_attr|
+    [:operation, :supplier, :code, :context, :happened_at].each do |required_attr|
       it "is a no-op in case parameter #{required_attr} is not given" do
         parameters.delete(required_attr)